+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Features
 
 ### Enhancements
+* Added client-side filtering in experiment list by `type` and `status`, in addition to `id` (GUID). ([#686](https://github.com/opensearch-project/dashboards-search-relevance/pull/686))
 
 ### Bug Fixes